% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_neighbour.R
\name{add_neighbour}
\alias{add_neighbour}
\title{Add isolated areas (polygons) to its nearest neighbour}
\usage{
add_neighbour(carto, nb = NULL, plot = FALSE)
}
\arguments{
\item{carto}{object of class \code{SpatialPolygonsDataFrame} or \code{sf}.}

\item{nb}{optional argument with the neighbours list of class \code{nb}.
If \code{NULL} (default), this object is computed from the \code{carto} argument.}

\item{plot}{logical value (default \code{FALSE}), if \code{TRUE} then the computed neighbourhood graph is ploted.}
}
\value{
This function returns a list with the following two elements:
\itemize{
  \item \code{nb}: the modified neighbours list
  \item \code{W}: associated spatial adjacency matrix of class \code{CsparseMatrix}
}
}
\description{
The function returns a neighbours list of class \code{nb} and its associated spatial adjacency matrix
computed by adding to isolated areas its nearest neighbour using the \code{knearneigh} function of 'spdep' package.
}
\examples{
library(spdep)

## load the Spain colorectal cancer mortality data ##
data(Carto_SpainMUN)

## compute the neighbours list from spatial polygons ##
nb_SpainMUN <- poly2nb(Carto_SpainMUN)
summary(nb_SpainMUN) # 1 region with no links

## add isolated area to its nearest neighbour ####
carto.mod <- add_neighbour(carto=Carto_SpainMUN, nb=nb_SpainMUN)
summary(carto.mod$nb) # 0 region with no links

}
