% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Carto_SpainMUN}
\alias{Carto_SpainMUN}
\title{Spanish colorectal cancer mortality data}
\format{
Formal class \code{sf}; the data contains a data.frame with 7907 rows and 11 variables.
\itemize{
  \item ID: character vector of geographic identifiers
  \item name: character vector of municipality names
  \item area: municipality polygon areas (in square meters)
  \item perimeter: municipality polygon perimeters (in meters)
  \item obs: observed number of cases
  \item exp: expected number of cases
  \item SMR: standardized mortality ratios
  \item region: character vector of autonomous regions
  \item geometry: sfc_MULTIPOLYGON
}
}
\usage{
Carto_SpainMUN
}
\description{
\code{sf} object containing the polygons of the municipalities of continental Spain
and simulated colorectal cancer mortality data.
}
\keyword{data}
