% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_LungCancer}
\alias{Data_LungCancer}
\title{Spanish lung cancer mortality data}
\format{
Formal class \code{data.frame} with 197.675 rows and 5 colunmns.
\itemize{
  \item ID: character vector of geographic identifiers
  \item year: numeric vector of year's identifiers
  \item obs: observed number of cases
  \item exp: expected number of cases
  \item SMR: standardized mortality ratios
  \item pop: population at risk
}
}
\usage{
Data_LungCancer
}
\description{
\code{data.frame} object containing simulated lung cancer mortality data in the 7907 municipalities of continental Spain during the period 1991-2015.
}
\keyword{data}
