% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_MultiCancer}
\alias{Data_MultiCancer}
\title{Spanish cancer mortality data for the joint analysis of multiple diseases}
\format{
Formal class \code{data.frame} with 237.271 rows and 5 colunmns.
\itemize{
  \item ID: character vector of geographic identifiers
  \item disease: numeric vector of disease identifiers
  \item obs: observed number of cases
  \item exp: expected number of cases
  \item SMR: standardized mortality ratios
}
}
\usage{
Data_MultiCancer
}
\description{
\code{data.frame} object containing simulated cancer mortality data for three diseases in the 7907 municipalities of continental Spain.
}
\keyword{data}
