% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mmodel_bym2.R
\name{Mmodel_bym2}
\alias{Mmodel_bym2}
\title{BYM2 multivariate CAR latent effect}
\usage{
Mmodel_bym2(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\arguments{
\item{cmd}{Internal functions used by the \code{rgeneric} model to define the latent effect.}

\item{theta}{Vector of hyperparameters.}
}
\value{
This is used internally by the \code{INLA::inla.rgeneric.define()} function.
}
\description{
M-model implementation of the BYM2 multivariate CAR latent effect using the \code{rgeneric} model of INLA.
}
\details{
This function considers an BYM2 CAR prior \insertCite{riebler2016intuitive}{bigDM} for the spatial latent effects of the different diseases and introduces correlation between them using the M-model proposal of \insertCite{botella2015unifying;textual}{bigDM}.
Putting the spatial latent effects for each disease in a matrix, the between disease dependence is introduced through the M matrix as \eqn{\Theta=\Phi M}, where the columns of \eqn{\Phi} follow an intrinsic BYM2 prior distribution (within-disease correlation).
A Wishart prior for the between covariance matrix \eqn{M'M} is considered using the Bartlett decomposition.
Uniform prior distributions on the interval [\code{alpha.min}, \code{alpha.max}] are considered for all the spatial mixing parameters.
The joint precision matrix of the spatial latent effects is given by \deqn{Q=(M^{-1}\otimes U)\operatorname{Blockdiag}(Q_1,\ldots,Q_J)(M^{-1}\otimes U)^\top,}
where \eqn{Q_j} denotes the diagonal BYM2 precision matrix in the spectral basis of the intrinsic CAR precision matrix, with \eqn{Q_{\mathrm{iCAR}}=U\Lambda U^\top}.
Since the resulting precision matrix is not sparse, this formulation is computationally more demanding than the corresponding CAR models.
\cr\cr
The following arguments are required to be defined before calling the functions:
\itemize{
\item \code{W}: binary adjacency matrix of the spatial areal units
\item \code{J}: number of diseases
\item \code{initial.values}: initial values defined for the cells of the M-matrix
\item \code{alpha.min}: lower limit defined for the uniform prior distribution of the spatial smoothing parameters
\item \code{alpha.max}: upper limit defined for the uniform prior distribution of the spatial smoothing parameters
}
}
\references{
\insertRef{botella2015unifying}{bigDM}

\insertRef{riebler2016intuitive}{bigDM}
}
\seealso{
\code{\link{Mmodel_icar}}, \code{\link{Mmodel_lcar}} and \code{\link{Mmodel_pcar}} for alternative multivariate CAR prior distributions.
}
