% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partition.R
\name{random_partition}
\alias{random_partition}
\title{Deprecated alias for \code{grid_partition()}}
\usage{
random_partition(...)
}
\arguments{
\item{...}{Arguments passed to \code{grid_partition()}.}
}
\description{
Deprecated alias for \code{grid_partition()}
}
