% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAR_INLA.R
\name{CAR_INLA}
\alias{CAR_INLA}
\title{CAR_INLA function}
\usage{
CAR_INLA(
  carto = NULL,
  ID.area = NULL,
  ID.group = NULL,
  O = NULL,
  E = NULL,
  prior = "Leroux",
  model = "partition",
  k = 0,
  strategy = "simplified.laplace",
  PCpriors = FALSE,
  seed = NULL,
  n.sample = 1000,
  compute.fixed = FALSE,
  compute.DIC = TRUE
)
}
\arguments{
\item{carto}{object of class \code{SpatialPolygonsDataFrame} or \code{sf} which, at least,
must contain the target variables of interest specified in the arguments \code{ID.area}, \code{O} and \code{E}.}

\item{ID.area}{character; name of the variable which contains the IDs of spatial areal units.}

\item{ID.group}{character; name of the variable which contains the IDs of the spatial partition (grouping variable).
Only required if \code{model="partition"}.}

\item{O}{character; name of the variable which contains the observed number of disease cases for each areal units.}

\item{E}{character; name of the variable which contains either the expected number of disease cases or the population at risk for each areal unit.}

\item{prior}{one of either "Leroux" (default), "intrinsic", "BYM" or "BYM2",
which specifies the prior distribution considered for the spatial random effect.}

\item{model}{one of either "global" or "partition" (default), which respectively specify the \emph{Global model}
or one of the scalable model proposal's (\emph{Disjoint model} and \emph{k-order neighbourhood model}).}

\item{k}{numeric value with the neighbourhood order used for the partition model.
If k=0 (default) the \emph{Disjoint model} is considered. Only required if \code{model="partition"}.}

\item{strategy}{one of either "gaussian", "simplified.laplace" (default), "laplace" or "adaptive",
which specifies the approximation strategy considered in the \code{inla} function.}

\item{PCpriors}{logical value (default \code{FALSE}); if \code{TRUE} then penalised complexity (PC) priors area used for the precision parameter of the spatial random effect.
Only works if \code{prior="intrinsic"} or \code{prior="BYM2"} arguments are specified.}

\item{seed}{numeric; control the RNG of \code{inla.qsample} (see \code{help(inla.qsample)} for further information). Defaults to \code{NULL}.}

\item{n.sample}{numeric; number of samples to generate from the posterior marginal distribution of the risks. Default to 1000.}

\item{compute.fixed}{logical value (default \code{FALSE}); if \code{TRUE} then the overall log-risk \eqn{\alpha} is computed.
Only works if \code{k=0} argument (\emph{disjoint model}) is specified.}

\item{compute.DIC}{logical value; if \code{TRUE} (default) then approximate values of the Deviance Information Criterion (DIC) and Watanabe-Akaike Information Criterion (WAIC) are computed.}
}
\value{
This function returns an object of class \code{inla}. See the \code{\link{mergeINLA}} function for details.
}
\description{
Description.
}
\details{
Details.
}
\examples{
## load the Spain colorectal cancer mortality data ##
data(Carto_SpainMUN)

\dontrun{
## fit the global model with a Leroux CAR prior distribution ##
Global <- CAR_INLA(carto=Carto_SpainMUN, ID.area="ID", O="obs", E="exp",
                   prior="Leroux", model="global", strategy="gaussian")

summary(Global)

## fit the disjoint model with a Leroux CAR prior distribution ##
Disjoint <- CAR_INLA(carto=Carto_SpainMUN, ID.area="ID", ID.group="region", O="obs", E="exp",
                     prior="Leroux", model="partition", k=0, strategy="gaussian")
summary(Disjoint)

## fit the 1st order neighbourhood model with a Leroux CAR prior distribution ##
order1 <- CAR_INLA(carto=Carto_SpainMUN, ID.area="ID", ID.group="region", O="obs", E="exp",
                   prior="Leroux", model="partition", k=1, strategy="gaussian")
summary(order1)

## fit the 2nd order neighbourhood model with a Leroux CAR prior distribution ##
order2 <- CAR_INLA(carto=Carto_SpainMUN, ID.area="ID", ID.group="region", O="obs", E="exp",
                   prior="Leroux", model="partition", k=2, strategy="gaussian")
summary(order2)
}

}
