% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_partition.R
\name{random_partition}
\alias{random_partition}
\title{Define a random partition of the spatial domain based on a regular grid}
\usage{
random_partition(carto, rows = 3, columns = 3)
}
\arguments{
\item{carto}{object of class \code{SpatialPolygonsDataFrame} or \code{sf}.}

\item{rows}{numeric; number of rows to define the regular grid. Default to 3.}

\item{columns}{numeric; number of columns to define the regular grid. Default to 3.}
}
\value{
\code{sf} object with the original data and a grouping variable named 'ID.group'
}
\description{
The function takes an object of class \code{SpatialPolygonsDataFrame} or \code{sf} and
defines a random partition of the spatial polygons based on a regular grid over the whole domain
using the \code{st_make_grid} function of 'sf' package.
}
\examples{
library(tmap)

## load the Spain colorectal cancer mortality data ##
data(Carto_SpainMUN)

## define a random partition based on a 3x3 regular grid ##
carto.new <- random_partition(carto=Carto_SpainMUN, rows=3, columns=3)

## plot of the grouping variable 'ID.group' ##
tm_shape(carto.new) +
  tm_polygons(col="ID.group") +
  tm_layout(legend.outside=TRUE)

}
